Remove duplicate dependencies by patching, and add back links field
We don't want duplicate ctru-sys crates in the dependency graph. Same
for libc. We can patch them here to avoid any issues.

Now that we don't have duplicate ctru-sys crates, we can add back the
links field. This field is important because it warns us when there are
two crates linking to the same library (ex. duplicate ctru-sys crates).