Fix link error with multiply-defined stack size
We don't want to define __stacksize__ more than once, since the linker
will reject it even if the definition is the same.

For some reason this seems to link fine with devkitARM release 62, but
not 61 (which is the version in the latest upstream docker images). This
is probably safer anyway, since there are technically two copies of
`ctru-rs` at play here.